+++ lib/ui/send/send_confirm_sheet.dart
@@ -16,6 +16,7 @@ import 'package:wallet_flutter/model/db/appdb.dart';
 import 'package:wallet_flutter/model/db/scheduled.dart';
 import 'package:wallet_flutter/model/db/txdata.dart';
 import 'package:wallet_flutter/model/db/user.dart';
+import 'package:wallet_flutter/model/db/work_source.dart';
 import 'package:wallet_flutter/model/vault.dart';
 import 'package:wallet_flutter/network/account_service.dart';
 import 'package:wallet_flutter/network/giftcards.dart';
@@ -83,7 +84,9 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
   StreamSubscription<AuthenticatedEvent>? _authSub;
 
   void _registerBus() {
-    _authSub = EventTaxiImpl.singleton().registerTo<AuthenticatedEvent>().listen((AuthenticatedEvent event) {
+    _authSub = EventTaxiImpl.singleton()
+        .registerTo<AuthenticatedEvent>()
+        .listen((AuthenticatedEvent event) {
       if (event.authType == AUTH_EVENT_TYPE.SEND) {
         _doSend();
       }
@@ -131,13 +134,15 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
 
   void _showAnimation(BuildContext context, AnimationType type) {
     animationOpen = true;
-    AppAnimation.animationLauncher(context, type, onPoppedCallback: () => animationOpen = false);
+    AppAnimation.animationLauncher(context, type,
+        onPoppedCallback: () => animationOpen = false);
   }
 
   @override
   Widget build(BuildContext context) {
     return SafeArea(
-        minimum: EdgeInsets.only(bottom: MediaQuery.of(context).size.height * 0.035),
+        minimum:
+            EdgeInsets.only(bottom: MediaQuery.of(context).size.height * 0.035),
         child: Column(
           children: <Widget>[
             Handlebars.horizontal(context),
@@ -153,7 +158,10 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                       children: <Widget>[
                         Text(
                           CaseChange.toUpperCase(
-                              (widget.link.isEmpty) ? Z.of(context).sending : Z.of(context).creatingGiftCard, context),
+                              (widget.link.isEmpty)
+                                  ? Z.of(context).sending
+                                  : Z.of(context).creatingGiftCard,
+                              context),
                           style: AppStyles.textStyleHeader(context),
                         ),
                       ],
@@ -162,13 +170,16 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                   // Container for the amount text
                   if (widget.memo.isNotEmpty && (widget.amountRaw == "0"))
                     Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 25.0, vertical: 15.0),
+                        padding: const EdgeInsets.symmetric(
+                            horizontal: 25.0, vertical: 15.0),
                         margin: EdgeInsets.only(
                             left: MediaQuery.of(context).size.width * 0.105,
                             right: MediaQuery.of(context).size.width * 0.105),
                         width: double.infinity,
                         decoration: BoxDecoration(
-                          color: StateContainer.of(context).curTheme.backgroundDarkest,
+                          color: StateContainer.of(context)
+                              .curTheme
+                              .backgroundDarkest,
                           borderRadius: BorderRadius.circular(25),
                         ),
                         child: Text(
@@ -181,10 +192,13 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                       margin: EdgeInsets.only(
                           left: MediaQuery.of(context).size.width * 0.105,
                           right: MediaQuery.of(context).size.width * 0.105),
-                      padding: const EdgeInsets.symmetric(horizontal: 25, vertical: 15),
+                      padding: const EdgeInsets.symmetric(
+                          horizontal: 25, vertical: 15),
                       width: double.infinity,
                       decoration: BoxDecoration(
-                        color: StateContainer.of(context).curTheme.backgroundDarkest,
+                        color: StateContainer.of(context)
+                            .curTheme
+                            .backgroundDarkest,
                         borderRadius: BorderRadius.circular(50),
                       ),
                       // Amount text
@@ -194,21 +208,32 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                           text: "",
                           children: [
                             TextSpan(
-                              text: getThemeAwareRawAccuracy(context, widget.amountRaw),
-                              style: AppStyles.textStyleParagraphPrimary(context),
+                              text: getThemeAwareRawAccuracy(
+                                  context, widget.amountRaw),
+                              style:
+                                  AppStyles.textStyleParagraphPrimary(context),
                             ),
                             displayCurrencySymbol(
                               context,
                               AppStyles.textStyleParagraphPrimary(context),
                             ),
                             TextSpan(
-                              text: getRawAsThemeAwareFormattedAmount(context, widget.amountRaw),
-                              style: AppStyles.textStyleParagraphPrimary(context),
+                              text: getRawAsThemeAwareFormattedAmount(
+                                  context, widget.amountRaw),
+                              style:
+                                  AppStyles.textStyleParagraphPrimary(context),
                             ),
                             TextSpan(
-                              text: widget.localCurrency != null ? " (${widget.localCurrency})" : "",
-                              style: AppStyles.textStyleParagraphPrimary(context).copyWith(
-                                color: StateContainer.of(context).curTheme.primary!.withOpacity(0.75),
+                              text: widget.localCurrency != null
+                                  ? " (${widget.localCurrency})"
+                                  : "",
+                              style:
+                                  AppStyles.textStyleParagraphPrimary(context)
+                                      .copyWith(
+                                color: StateContainer.of(context)
+                                    .curTheme
+                                    .primary!
+                                    .withOpacity(0.75),
                               ),
                             ),
                           ],
@@ -232,17 +257,21 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                   // Address text
                   if (widget.link.isEmpty)
                     Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 25.0, vertical: 15.0),
+                        padding: const EdgeInsets.symmetric(
+                            horizontal: 25.0, vertical: 15.0),
                         margin: EdgeInsets.only(
                             left: MediaQuery.of(context).size.width * 0.105,
                             right: MediaQuery.of(context).size.width * 0.105),
                         width: double.infinity,
                         decoration: BoxDecoration(
-                          color: StateContainer.of(context).curTheme.backgroundDarkest,
+                          color: StateContainer.of(context)
+                              .curTheme
+                              .backgroundDarkest,
                           borderRadius: BorderRadius.circular(25),
                         ),
-                        child:
-                            UIUtil.threeLineAddressText(context, widget.destination, contactName: widget.contactName)),
+                        child: UIUtil.threeLineAddressText(
+                            context, widget.destination,
+                            contactName: widget.contactName)),
 
                   // WITH MESSAGE:
                   if (widget.memo.isNotEmpty && (widget.amountRaw != "0"))
@@ -251,7 +280,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                       child: Column(
                         children: <Widget>[
                           Text(
-                            CaseChange.toUpperCase(Z.of(context).withMessage, context),
+                            CaseChange.toUpperCase(
+                                Z.of(context).withMessage, context),
                             style: AppStyles.textStyleHeader(context),
                           ),
                         ],
@@ -260,13 +290,16 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                   // MEMO:
                   if (widget.memo.isNotEmpty && (widget.amountRaw != "0"))
                     Container(
-                        padding: const EdgeInsets.symmetric(horizontal: 25.0, vertical: 15.0),
+                        padding: const EdgeInsets.symmetric(
+                            horizontal: 25.0, vertical: 15.0),
                         margin: EdgeInsets.only(
                             left: MediaQuery.of(context).size.width * 0.105,
                             right: MediaQuery.of(context).size.width * 0.105),
                         width: double.infinity,
                         decoration: BoxDecoration(
-                          color: StateContainer.of(context).curTheme.backgroundDarkest,
+                          color: StateContainer.of(context)
+                              .curTheme
+                              .backgroundDarkest,
                           borderRadius: BorderRadius.circular(25),
                         ),
                         child: Text(
@@ -334,16 +367,22 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                       Navigator.of(context).pop();
                     }),
                     // CONFIRM Button
-                    AppButton.buildAppButton(context, AppButtonType.PRIMARY,
-                        CaseChange.toUpperCase(Z.of(context).confirm, context), Dimens.BUTTON_COMPACT_RIGHT_DIMENS,
+                    AppButton.buildAppButton(
+                        context,
+                        AppButtonType.PRIMARY,
+                        CaseChange.toUpperCase(Z.of(context).confirm, context),
+                        Dimens.BUTTON_COMPACT_RIGHT_DIMENS,
                         disabled: !shownWarning, onPressed: () async {
                       if (clicking) return;
                       clicking = true;
                       // Authenticate
-                      final AuthenticationMethod authMethod = await sl.get<SharedPrefsUtil>().getAuthMethod();
-                      final bool hasBiometrics = await sl.get<BiometricUtil>().hasBiometrics();
+                      final AuthenticationMethod authMethod =
+                          await sl.get<SharedPrefsUtil>().getAuthMethod();
+                      final bool hasBiometrics =
+                          await sl.get<BiometricUtil>().hasBiometrics();
 
-                      final bool isMessage = widget.memo.isNotEmpty && (widget.amountRaw == "0");
+                      final bool isMessage =
+                          widget.memo.isNotEmpty && (widget.amountRaw == "0");
 
                       if (!mounted) return;
 
@@ -352,18 +391,25 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                           : Z
                               .of(context)
                               .sendAmountConfirm
-                              .replaceAll("%1", getRawAsThemeAwareAmount(context, widget.amountRaw))
-                              .replaceAll("%2", StateContainer.of(context).currencyMode);
+                              .replaceAll(
+                                  "%1",
+                                  getRawAsThemeAwareAmount(
+                                      context, widget.amountRaw))
+                              .replaceAll("%2",
+                                  StateContainer.of(context).currencyMode);
 
                       if (!mounted) return;
 
-                      if (authMethod.method == AuthMethod.BIOMETRICS && hasBiometrics) {
+                      if (authMethod.method == AuthMethod.BIOMETRICS &&
+                          hasBiometrics) {
                         try {
-                          final bool authenticated =
-                              await sl.get<BiometricUtil>().authenticateWithBiometrics(context, authText);
+                          final bool authenticated = await sl
+                              .get<BiometricUtil>()
+                              .authenticateWithBiometrics(context, authText);
                           if (authenticated) {
                             sl.get<HapticUtil>().fingerprintSucess();
-                            EventTaxiImpl.singleton().fire(AuthenticatedEvent(AUTH_EVENT_TYPE.SEND));
+                            EventTaxiImpl.singleton()
+                                .fire(AuthenticatedEvent(AUTH_EVENT_TYPE.SEND));
                           }
                         } catch (e) {
                           await authenticateWithPin();
@@ -371,7 +417,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
                       } else if (authMethod.method == AuthMethod.PIN) {
                         await authenticateWithPin();
                       } else {
-                        EventTaxiImpl.singleton().fire(AuthenticatedEvent(AUTH_EVENT_TYPE.SEND));
+                        EventTaxiImpl.singleton()
+                            .fire(AuthenticatedEvent(AUTH_EVENT_TYPE.SEND));
                       }
                       clicking = false;
                     })
@@ -392,13 +439,15 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
 
   Future<bool> showUnopenedWarning(String address) async {
     // if we have the warn setting on, and the account isn't open, show the dialog:
-    final bool warningOn = await sl.get<SharedPrefsUtil>().getUnopenedWarningOn();
+    final bool warningOn =
+        await sl.get<SharedPrefsUtil>().getUnopenedWarningOn();
     if (!warningOn) {
       return true;
     }
     // check if the address is open:
     try {
-      final AccountInfoResponse accountInfo = await sl.get<AccountService>().getAccountInfo(address);
+      final AccountInfoResponse accountInfo =
+          await sl.get<AccountService>().getAccountInfo(address);
       if (!accountInfo.unopened) {
         return true;
       }
@@ -426,7 +475,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
               mainAxisSize: MainAxisSize.min,
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: <Widget>[
-                Text("${Z.of(context).unopenedWarningWarning}\n\n", style: AppStyles.textStyleParagraph(context)),
+                Text("${Z.of(context).unopenedWarningWarning}\n\n",
+                    style: AppStyles.textStyleParagraph(context)),
                 RichText(
                   textAlign: TextAlign.start,
                   text: TextSpan(
@@ -481,15 +531,31 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
       final bool isMessage = widget.amountRaw == "0";
       final String walletAddress = StateContainer.of(context).wallet!.address!;
 
-      _showAnimation(context, isMessage ? AnimationType.SEND_MESSAGE : AnimationType.SEND);
+      _showAnimation(
+          context, isMessage ? AnimationType.SEND_MESSAGE : AnimationType.SEND);
 
       ProcessResponse? resp;
 
-      final String derivationMethod = await sl.get<SharedPrefsUtil>().getKeyDerivationMethod();
+      final String derivationMethod =
+          await sl.get<SharedPrefsUtil>().getKeyDerivationMethod();
       if (!isMessage) {
         if (!obscuredMode) {
-          final String privKey = await NanoUtil.uniSeedToPrivate(await StateContainer.of(context).getSeed(),
-              StateContainer.of(context).selectedAccount!.index!, derivationMethod);
+          final String privKey = await NanoUtil.uniSeedToPrivate(
+              await StateContainer.of(context).getSeed(),
+              StateContainer.of(context).selectedAccount!.index!,
+              derivationMethod);
+
+          // check if using local work generation:
+          final WorkSource ws =
+              await sl.get<DBHelper>().getSelectedWorkSource();
+          if (ws.type == WorkSourceTypes.LOCAL) {
+            if (!mounted) return;
+            UIUtil.showSnackbar(
+              Z.of(context).generatingWork,
+              context,
+              durationMs: 25000,
+            );
+          }
           resp = await sl.get<AccountService>().requestSend(
                 StateContainer.of(context).wallet!.representative,
                 StateContainer.of(context).wallet!.frontier,
@@ -501,7 +567,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
               );
           if (!mounted) return;
           StateContainer.of(context).wallet!.frontier = resp.hash;
-          StateContainer.of(context).wallet!.accountBalance -= BigInt.parse(widget.amountRaw);
+          StateContainer.of(context).wallet!.accountBalance -=
+              BigInt.parse(widget.amountRaw);
         } else {
           sl.get<Logger>().v("OBSCURED MODE");
 
@@ -514,8 +581,10 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
             derivationMethod,
           );
 
-          final String privKey = await NanoUtil.uniSeedToPrivate(await StateContainer.of(context).getSeed(),
-              StateContainer.of(context).selectedAccount!.index!, derivationMethod);
+          final String privKey = await NanoUtil.uniSeedToPrivate(
+              await StateContainer.of(context).getSeed(),
+              StateContainer.of(context).selectedAccount!.index!,
+              derivationMethod);
           resp = await sl.get<AccountService>().requestSend(
                 StateContainer.of(context).wallet!.representative,
                 StateContainer.of(context).wallet!.frontier,
@@ -527,13 +596,17 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
               );
           if (!mounted) return;
           StateContainer.of(context).wallet!.frontier = resp.hash;
-          StateContainer.of(context).wallet!.accountBalance -= BigInt.parse(widget.amountRaw);
+          StateContainer.of(context).wallet!.accountBalance -=
+              BigInt.parse(widget.amountRaw);
 
           sl.get<Logger>().v("SENT TO ADDRESS 200");
 
           // receive from address 200:
-          await AppTransferOverviewSheetState()
-              .receiveAtIndex(context, await StateContainer.of(context).getSeed(), randomIndex, derivationMethod);
+          await AppTransferOverviewSheetState().receiveAtIndex(
+              context,
+              await StateContainer.of(context).getSeed(),
+              randomIndex,
+              derivationMethod);
 
           sl.get<Logger>().v("RECEIVED AT ADDRESS 200");
 
@@ -563,14 +636,15 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
           derivationMethod,
         );
         // get epoch time as hex:
-        final int secondsSinceEpoch = DateTime.now().millisecondsSinceEpoch ~/ Duration.millisecondsPerSecond;
+        final int secondsSinceEpoch = DateTime.now().millisecondsSinceEpoch ~/
+            Duration.millisecondsPerSecond;
         final String nonceHex = secondsSinceEpoch.toRadixString(16);
         final String signature = NanoSignatures.signBlock(nonceHex, privKey);
 
         // check validity locally:
         final String pubKey = NanoAccounts.extractPublicKey(walletAddress);
-        final bool isValid =
-            NanoSignatures.validateSig(nonceHex, NanoHelpers.hexToBytes(pubKey), NanoHelpers.hexToBytes(signature));
+        final bool isValid = NanoSignatures.validateSig(nonceHex,
+            NanoHelpers.hexToBytes(pubKey), NanoHelpers.hexToBytes(signature));
         if (!isValid) {
           throw Exception("Invalid signature?!");
         }
@@ -579,9 +653,10 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
         const Uuid uuid = Uuid();
         final String localUuid = "LOCAL:${uuid.v4()}";
         // current block height:
-        final int currentBlockHeightInList = StateContainer.of(context).wallet!.history.isNotEmpty
-            ? (StateContainer.of(context).wallet!.history[0].height! + 1)
-            : 1;
+        final int currentBlockHeightInList =
+            StateContainer.of(context).wallet!.history.isNotEmpty
+                ? (StateContainer.of(context).wallet!.history[0].height! + 1)
+                : 1;
         final TXData memoTXData = TXData(
           from_address: walletAddress,
           to_address: widget.destination,
@@ -602,16 +677,23 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
 
         try {
           // encrypt the memo:
-          final String encryptedMemo = Box.encrypt(widget.memo, widget.destination, privKey);
+          final String encryptedMemo =
+              Box.encrypt(widget.memo, widget.destination, privKey);
 
           if (isMessage) {
-            await sl
-                .get<MetadataService>()
-                .sendTXMessage(widget.destination, walletAddress, signature, nonceHex, encryptedMemo, localUuid);
+            await sl.get<MetadataService>().sendTXMessage(widget.destination,
+                walletAddress, signature, nonceHex, encryptedMemo, localUuid);
           } else {
             // just a memo:
-            await sl.get<MetadataService>().sendTXMemo(widget.destination, walletAddress, widget.amountRaw, signature,
-                nonceHex, encryptedMemo, resp?.hash, localUuid);
+            await sl.get<MetadataService>().sendTXMemo(
+                widget.destination,
+                walletAddress,
+                widget.amountRaw,
+                signature,
+                nonceHex,
+                encryptedMemo,
+                resp?.hash,
+                localUuid);
           }
         } catch (e) {
           sl.get<Logger>().v("error encrypting memo: $e");
@@ -636,7 +718,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
       }
 
       // go through and check to see if any unfulfilled payments are now fulfilled
-      final List<TXData> unfulfilledPayments = await sl.get<DBHelper>().getUnfulfilledTXs();
+      final List<TXData> unfulfilledPayments =
+          await sl.get<DBHelper>().getUnfulfilledTXs();
       for (int i = 0; i < unfulfilledPayments.length; i++) {
         final TXData txData = unfulfilledPayments[i];
 
@@ -662,11 +745,13 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
       }
 
       // check if this fulfilled any subscriptions / scheduled payments:
-      final List<Scheduled> scheduledPayments = await sl.get<DBHelper>().getScheduled();
+      final List<Scheduled> scheduledPayments =
+          await sl.get<DBHelper>().getScheduled();
       for (int i = 0; i < scheduledPayments.length; i++) {
         final Scheduled scheduled = scheduledPayments[i];
         // check to make sure the recipient is correct and the amount is correct:
-        if (scheduled.address == widget.destination && scheduled.amount_raw == widget.amountRaw) {
+        if (scheduled.address == widget.destination &&
+            scheduled.amount_raw == widget.amountRaw) {
           // make sure the payment was due:
           final int currentTime = DateTime.now().millisecondsSinceEpoch ~/ 1000;
           if (scheduled.timestamp < currentTime) {
@@ -678,7 +763,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
       // Show complete
       String? contactName = widget.contactName;
       if (widget.contactName == null || widget.contactName!.isEmpty) {
-        final User? user = await sl.get<DBHelper>().getUserWithAddress(widget.destination);
+        final User? user =
+            await sl.get<DBHelper>().getUserWithAddress(widget.destination);
         if (user != null) {
           contactName = user.getDisplayName();
         }
@@ -692,7 +778,12 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
 
       if (memoSendFailed) {
         Navigator.of(context).popUntil(RouteUtils.withNameLike('/home'));
-        UIUtil.showSnackbar(Z.of(context).sendMemoError.replaceAll("%1", NonTranslatable.appName), context,
+        UIUtil.showSnackbar(
+            Z
+                .of(context)
+                .sendMemoError
+                .replaceAll("%1", NonTranslatable.appName),
+            context,
             durationMs: 5000);
       } else {
         if (widget.link.isEmpty) {
@@ -735,7 +826,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
     // PIN Authentication
     final String? expectedPin = await sl.get<Vault>().getPin();
     final String? plausiblePin = await sl.get<Vault>().getPlausiblePin();
-    final bool? auth = await Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context) {
+    final bool? auth = await Navigator.of(context)
+        .push(MaterialPageRoute(builder: (BuildContext context) {
       return PinScreen(
         PinOverlayType.ENTER_PIN,
         expectedPin: expectedPin,
@@ -743,7 +835,8 @@ class _SendConfirmSheetState extends State<SendConfirmSheet> {
         description: Z
             .of(context)
             .sendAmountConfirm
-            .replaceAll("%1", getRawAsThemeAwareAmount(context, widget.amountRaw))
+            .replaceAll(
+                "%1", getRawAsThemeAwareAmount(context, widget.amountRaw))
             .replaceAll("%2", StateContainer.of(context).currencyMode),
       );
     }));